backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Async & Background Tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
channels==4.0.0
channels-redis==4.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-allauth==0.60.0
python-jose==3.3.0
cryptography==41.0.7

# AI & ML
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.22
tiktoken==0.5.2

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
openpyxl==3.1.2

# API & Serialization
pydantic==2.5.3
marshmallow==3.20.1

# Utilities
python-dateutil==2.8.2
requests==2.31.0
httpx==0.26.0
aiohttp==3.9.1

# Monitoring & Logging
sentry-sdk==1.39.1
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-asyncio==0.23.3
factory-boy==3.3.0
faker==22.0.0

# Production Server
gunicorn==21.2.0
uvicorn==0.25.0

# Development
ipython==8.19.0
django-debug-toolbar==4.2.0
